requirements_dev.txt
pytest>=6.0
pytest-cov>=2.0
mypy>=0.910
flake8>=3.9
tox>=3.24
numpy
scipy
scikit-learn
torch
tqdm
matplotlib
pandas
seaborn
pyyaml
tensorboard
h5py
ete3
biopython